Move OSS caching strategy definitions into the caching plugin
The plugins module imported strategies, getPerformanceTabMetadata, and
ModelWithClearableCache from admin/performance to use as PLUGIN_CACHING
defaults. This was also a circular dependency: plugins -> admin constants
-> admin utils -> PLUGIN_CACHING (plugins).

Invert the ownership: the OSS strategy definitions (and the types and
validation schemas they're built from) now live in the caching plugin,
which is their runtime source of truth - even admin's own utils read
strategies via PLUGIN_CACHING, and enterprise extends them through the
plugin. All admin and enterprise consumers now import these from
metabase/plugins directly. strategyValidationSchema and
getPerformanceAdminPaths stay in admin (form/navigation concerns).

Fixes 2 module boundary violations (shared/plugins -> feature/admin) and
removes the circular import.
