+++ enterprise/frontend/src/metabase-enterprise/caching/components/InvalidateNowButton.tsx
@@ -5,10 +5,12 @@ import { c, t } from "ttag";
 
 import { useInvalidateTarget } from "metabase/admin/performance/hooks/useInvalidateTarget";
 import { useIsFormPending } from "metabase/admin/performance/hooks/useIsFormPending";
-import type { ModelWithClearableCache } from "metabase/admin/performance/types";
 import { useConfirmation } from "metabase/common/hooks/use-confirmation";
 import { FormProvider, FormSubmitButton } from "metabase/forms";
-import type { InvalidateNowButtonProps } from "metabase/plugins";
+import type {
+  InvalidateNowButtonProps,
+  ModelWithClearableCache,
+} from "metabase/plugins";
 import { Group, Icon, Loader, Text } from "metabase/ui";
 
 /** Button that clears the cache of a particular object (the "target") */
