+++ frontend/src/metabase/admin/performance/constants/complex.ts
@@ -1,90 +1,6 @@
 import { t } from "ttag";
-import * as Yup from "yup";
 
 import type { AdminPath } from "metabase/redux/store";
-import type { CacheableModel } from "metabase-types/api";
-
-import { PerformanceTabId, type StrategyData } from "../types";
-
-import { defaultMinDurationMs } from "./simple";
-
-/** Rather than a constant defined in the module scope, this is a function. This way, ttag.t runs *after* the locale is set */
-export const getPositiveIntegerSchema = () =>
-  Yup.number()
-    .positive(t`Enter a positive number.`)
-    .integer(t`Enter an integer.`);
-
-export const inheritStrategyValidationSchema = Yup.object({
-  type: Yup.string().equals(["inherit"]),
-});
-
-export const doNotCacheStrategyValidationSchema = Yup.object({
-  type: Yup.string().equals(["nocache"]),
-});
-
-/** Rather than a constant defined in the module scope, this is a function. This way, ttag.t runs *after* the locale is set */
-export const getAdaptiveStrategyValidationSchema = () => {
-  const positiveInteger = getPositiveIntegerSchema();
-  return Yup.object({
-    type: Yup.string().equals(["ttl"]),
-    min_duration_ms: positiveInteger.default(defaultMinDurationMs),
-    min_duration_seconds: positiveInteger.default(
-      Math.ceil(defaultMinDurationMs / 1000),
-    ),
-    multiplier: positiveInteger.default(10),
-  });
-};
-
-export const strategies = {
-  inherit: {
-    // NOTE: We use functions for labels because otherwise t doesn't work properly
-    label: () => t`Default`,
-    description: (model?: CacheableModel) => {
-      switch (model) {
-        case "database":
-          return t`Use the default policy`;
-        case "dashboard":
-          return t`Each question will use its own policy or the database policy`;
-        case "question":
-        case "metric":
-          return t`Use the database or dashboard policy`;
-        default:
-          return undefined;
-      }
-    },
-    shortLabel: () => t`Default`,
-    validationSchema: inheritStrategyValidationSchema,
-  },
-  // NOTE: The strategy is called 'ttl' in the BE, but we've renamed it 'Adaptive' in the FE
-  ttl: {
-    label: () => t`Adaptive`,
-    description: () =>
-      t`Use a query’s average execution time to determine how long to cache its results`,
-    shortLabel: () => t`Adaptive`,
-    validationSchema: getAdaptiveStrategyValidationSchema,
-  },
-  nocache: {
-    label: () => t`Don’t cache results`,
-    shortLabel: () => t`No caching`,
-    validationSchema: doNotCacheStrategyValidationSchema,
-  },
-} as Record<string, StrategyData>;
-
-export const getPerformanceTabMetadata = () =>
-  [
-    {
-      name: t`Database caching`,
-      path: "/admin/performance/databases",
-      key: "performance-databases",
-      tabId: PerformanceTabId.Databases,
-    },
-    {
-      name: t`Model persistence`,
-      path: "/admin/performance/models",
-      key: "performance-models",
-      tabId: PerformanceTabId.Models,
-    },
-  ] as (AdminPath & { tabId: string })[];
 
 export const getPerformanceAdminPaths = (metadata: AdminPath[]) =>
   metadata.map((tab) => ({ ...tab, name: `${t`Performance`} - ${tab.name}` }));
