+++ frontend/src/metabase/admin/performance/constants/simple.ts
@@ -1,3 +1,2 @@
 export const rootId = 0;
-export const defaultMinDurationMs = 1000;
 export const defaultCronSchedule = "0 0 * * * ? *";
