+++ frontend/src/metabase/admin/performance/utils.tsx
@@ -5,7 +5,14 @@ import type { SchemaObjectDescription } from "yup/lib/schema";
 
 import { cronToScheduleSettings } from "metabase/common/components/Schedule/cron";
 import { getScheduleStrings } from "metabase/common/components/Schedule/strings";
-import { PLUGIN_CACHING } from "metabase/plugins";
+import {
+  PLUGIN_CACHING,
+  type PerformanceTabId,
+  type StrategyData,
+  type StrategyLabel,
+  defaultMinDurationMs,
+  strategies,
+} from "metabase/plugins";
 import { isNullOrUndefined } from "metabase/utils/types";
 import type {
   AdaptiveStrategy,
@@ -15,9 +22,7 @@ import type {
   CacheableModel,
 } from "metabase-types/api";
 
-import { strategies } from "./constants/complex";
-import { defaultMinDurationMs, rootId } from "./constants/simple";
-import type { PerformanceTabId, StrategyData, StrategyLabel } from "./types";
+import { rootId } from "./constants/simple";
 
 type ErrorWithMessage = { data: { message: string } };
 export const isErrorWithMessage = (error: unknown): error is ErrorWithMessage =>
